fix: Revert breaking rename into `sls_sdk`
This reverts commit a15ba5f684936cc7063d09d9301105b6db240638.
