+++ python/packages/sdk/serverless_sdk/lib/stack_trace_string.py
