+++ python/packages/sdk/serverless_sdk/lib/warning_captured_event.py
