+++ python/packages/sdk/tests/lib/test_timing.py
@@ -1,4 +1,4 @@
-from sls_sdk.lib.timing import to_protobuf_epoch_timestamp
+from serverless_sdk.lib.timing import to_protobuf_epoch_timestamp
 import time
